renaming files CamelCase
