+++ src/const.ts
@@ -1,0 +1,3 @@
+export const SETTING = {
+  offersCount: 312
+};
