--- src/const.tsx
@@ -1,3 +1,0 @@
-export const Setting = {
-  offersCount: 312
-};
