+++ src/pages/FavoritesPage/FavoritesPage.tsx
